# perf/perf_pmu: Handle non root user creation