@sacsant 2023-06-27
Before fix:

 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: ERROR: Command 'useradd -g test_pmu -m -d /home/test_pmu test_pmu' failed.\nstdout: b''\nstderr: b'useradd: existing lock file /etc/subgid.lock without a PID\nuseradd: cannot lock /etc/subgid; try again later.\n'\nadditional_info: None (0.06 s)
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: STARTED
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: ERROR: Command 'useradd -g test_pmu -m -d /home/test_pmu test_pmu' failed.\nstdout: b''\nstderr: b'useradd: existing lock file /etc/subgid.lock without a PID\nuseradd: cannot lock /etc/subgid; try again later.\n'\nadditional_info: None (0.07 s)
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: STARTED
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: PASS (0.03 s)
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: STARTED
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: PASS (0.03 s)
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: STARTED
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: SKIP: This test is for PowerNV
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: STARTED
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: SKIP: This test is for PowerNV
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: STARTED
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: SKIP: This test is for PowerNV
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: STARTED
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: SKIP: This test is for PowerNV
RESULTS    : PASS 2 | ERROR 2 | FAIL 0 | SKIP 4 | WARN 0 | INTERRUPT 0 | CANCEL 0

After fix:

 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
 (1/8) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: WARN: Test passed but there were warnings during execution. Check the log for details. (0.05 s)
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: STARTED
 (2/8) perf/perf_pmu.py:PerfBasic.test_config_PPC_RTAS: WARN: Test passed but there were warnings during execution. Check the log for details. (0.05 s)
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: STARTED
 (3/8) perf/perf_pmu.py:PerfBasic.test_cpu_event_count: PASS (0.03 s)
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: STARTED
 (4/8) perf/perf_pmu.py:PerfBasic.test_hv_24x7_event_count: PASS (0.03 s)
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: STARTED
 (5/8) perf/perf_pmu.py:PerfBasic.test_imc_event_count: SKIP: This test is for PowerNV
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: STARTED
 (6/8) perf/perf_pmu.py:PerfBasic.test_thread_event_count: SKIP: This test is for PowerNV
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: STARTED
 (7/8) perf/perf_pmu.py:PerfBasic.test_nest_cpumask: SKIP: This test is for PowerNV
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: STARTED
 (8/8) perf/perf_pmu.py:PerfBasic.test_core_cpumask: SKIP: This test is for PowerNV
RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 4 | WARN 2 | INTERRUPT 0 | CANCEL 0